Use also error.rs with vm-service similar to the rest
